Added new endpoints to sleepData.
getDataByDay: get all the data in a day from a user.
getdataByWeek: get the sleep data in a week from a user.
getDataByMonth: get the sleep data in a month from a user.
getAverageWeekSleepData: get the average of hours of sleep in a week for a user
getAvergeMonthSleepData: get the averge of hours of sleep in a month for a user

In all of these endpoints, the body of the request must have userId and day.
Internally, the week and the month are calculated.